# internal/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Go microservice template for Kubernetes.",
        "title": "Bricks Backend POC API",
        "contact": {
            "name": "Source Code",
            "url": "https://github.com/nacuellar25111992/bricks-backend-poc"
        },
        "license": {
            "name": "MIT License",
            "url": "https://github.com/nacuellar25111992/bricks-backend-poc/blob/master/LICENSE"
        },
        "version": "2.0"
    },
    "host": "localhost:9898",
    "basePath": "/",
    "paths": {
...
        "/token/validate": {
            "post": {
                "description": "validates the JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "HTTP API"
                ],
                "summary": "Validate JWT token",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TokenValidationResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.TokenValidationResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "token_name": {
                    "type": "string"
                }
            }
        }
    }
}